==> scripts/3_QC_stats_Unfilt_Filt.Rmd <==
---
title: "QC stats"
author: "Yuvarani Masarapu"
date: "4/9/2021"
output: html_document
---

```{r global-options, include=FALSE}
knitr::opts_chunk$set(fig.path='figs/', fig.width=14, fig.height=10, warning=FALSE, message=FALSE, echo = FALSE)
```

```{r libraries-load, echo=FALSE}
suppressPackageStartupMessages({
  library(STutility)
  library(Seurat)
  library(harmony)
  library(gridExtra)
  library(pals)
  library(akima)
  library(readr)
  library(plotly)
  library(tibble)
  library(raster)
  library(dplyr)
  library(stringr)
  library(magrittr)
  library(ggplot2)
  library(imager)
  library(Matrix)
  library(cowplot)
  require(data.table)
  library(magick)
  library(grid)
  library(SeuratObject)
})
```

# General statistics and plots

### Statistics on Raw data (Unfiltered, Untransformed, not Normalised)

Violin plot showing distribution of gene counts and UMI counts.

```{r violin-unflt-data, message=FALSE, warning=FALSE, echo=FALSE}
spruce.unfilt <- readRDS("~/Documents/23-02-2021_spruce_new/210222_spruce_raw_data.rds")

VlnPlot(spruce.unfilt, c("nCount_RNA","nFeature_RNA"), pt.size = 0.1)
```

Histogram to plot count/expression distribution (raw data).

```{r counts-unfiltt-data, message=FALSE, warning=FALSE, echo=FALSE}
p1 <- ggplot() +
  geom_histogram(data = spruce.unfilt[[]], aes(nFeature_RNA), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  ggtitle("Unique genes per spot")

p2 <- ggplot() +
  geom_histogram(data = spruce.unfilt[[]], aes(nCount_RNA), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  ggtitle("Total counts per spots")

gene_attr <- data.frame(nUMI = Matrix::rowSums(spruce.unfilt@assays$RNA@counts), 
                        nSpots = Matrix::rowSums(spruce.unfilt@assays$RNA@counts > 0))
p3 <- ggplot() +
  geom_histogram(data = gene_attr, aes(nUMI), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  scale_x_log10() +
  ggtitle("Total counts per gene (log10 scale)")

p4 <- ggplot() +
  geom_histogram(data = gene_attr, aes(nSpots), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  ggtitle("Total spots per gene")

cowplot::plot_grid(p1, p2, p3, p4)
```

Now, we Run PCA on 'untransformed' unfiltered data, find clusters with PCA embeddings

```{r PCA-unfilt-data, message=FALSE, warning=FALSE, echo=FALSE }
spruce.unfiltered.PCA <- ScaleData(spruce.unfilt) %>% FindVariableFeatures() %>% RunPCA() %>% RunUMAP(reduction = "pca", dims = 1:30) %>% FindNeighbors(reduction = "pca", dims = 1:30) %>% FindClusters()

saveRDS(spruce.unfiltered.PCA, "spruce_unfiltered_PCA.rds")
```


###Statistics for filtered, normalised data

Violin plots showing distribution of gene counts and UMI counts.

```{r violin-SCT-data, message=FALSE, warning=FALSE, echo=FALSE }
spruce.filt.SCT <- readRDS("~/Documents/23-02-2021_spruce_new/spruce_merged_obj_19-03-2021.rds")

VlnPlot(spruce.filt.SCT, c("nCount_RNA","nFeature_RNA"), pt.size = 0.1)
```


Histogram to plot count/expression distribution.

```{r plots-SCT-data, message=FALSE, warning=FALSE, echo=FALSE}
p1 <- ggplot() +
  geom_histogram(data = spruce.filt.SCT[[]], aes(nFeature_RNA), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  ggtitle("Unique genes per spot")

p2 <- ggplot() +
  geom_histogram(data = spruce.filt.SCT[[]], aes(nCount_RNA), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  ggtitle("Total counts per spots")

gene_attr <- data.frame(nUMI = Matrix::rowSums(spruce.filt.SCT@assays$RNA@counts), 
                        nSpots = Matrix::rowSums(spruce.filt.SCT@assays$RNA@counts > 0))
p3 <- ggplot() +
  geom_histogram(data = gene_attr, aes(nUMI), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  scale_x_log10() +
  ggtitle("Total counts per gene (log10 scale)")

p4 <- ggplot() +
  geom_histogram(data = gene_attr, aes(nSpots), fill = "red", alpha = 0.7, color = "gray", bins = 50) +
  Seurat::DarkTheme() +
  ggtitle("Total spots per gene")

cowplot::plot_grid(p1, p2, p3, p4)
```
